Move around the home subpage files
